+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 * **ChatMessages:** allow message props to override role defaults
 * **useEditorMenu:** rank filtered results by relevance
 * **NavigationMenu:** streamline linkLabelExternalIcon rendering by nesting component into linkLabel
+* **Skeleton:** improve colors
 
 ## [2.2.1](https://github.com/bitrix24/b24ui/compare/v2.1.17...v2.2.1) (2025-12-18)